readded imgui

+++ IceSDK/Graphics/Components/AnimationComponent.h
@@ -6,8 +6,8 @@
 namespace IceSDK::Graphics::Components {
     struct AnimationComponent { // basic animation component
         std::vector<glm::vec2> frames;
-        float delay;
-        float last_frame;
+        double delay;
+        double last_frame;
         uint32_t current_frame;
     };
 } // namespace IceSDK::Graphics::Components
